Refactor BlockEdit into modular components
